Adjust URL paths to IMDB databases
Drat, it looks like it will soon require an S3 account to download the
IMDB data files. Should I pivot to requiring every student to create an
Amazon AWS account and giving them their credit card? For the moment, I
am adjusting the URLs to refer to files that seem not to return 550
errors. Fixes #32.